requirements.txt
oracledb
python-dotenv
tabulate
pytest
